requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
authlib==1.3.0
pyjwt[crypto]==2.8.0
python3-saml==1.16.0
python-dotenv==1.0.1
jinja2==3.1.3
httpx==0.27.0
lxml==5.1.0
beautifulsoup4==4.12.3
python-jose[cryptography]==3.3.0
pytest==8.1.1
itsdangerous==2.1.2
python-multipart==0.0.9
